pyproject.toml
[project]
name = "nds1"
version = "0.1.0"
description = "nds1"
readme = "README.md"
requires-python = ">=3.10, <4.0"
dependencies = [
    "nonebot2[fastapi]>=2.4.4",
    "nonebot-adapter-onebot>=2.4.6",
    "python-dotenv>=1.0.0",
    "anthropic>=0.40.0",
]

[tool.nonebot]
plugin_dirs = ["plugins"]
builtin_plugins = ["echo"]


[tool.nonebot.adapters]
nonebot-adapter-onebot = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
"@local" = []

[tool.nonebot.plugins]
"@local" = []

